Feat: Integration Roles. At Current Project

+++ ControlReparacions/app/Controllers/TicketsController.php
@@ -75,36 +75,87 @@ class TicketsController extends BaseController
             'search' => $search,
             'table' => $table,
         ];
-
+        $role = session()->get('user')['role'];
+        $institute = session()->get('user')['code'];
+        
+        // dd(session()->get('user'));
         // ROWS
         foreach ($data['tickets'] as $ticket) {
-
+            
             $buttonDelete = base_url("tickets/delete/" . $ticket['id']);
             $buttonUpdate = base_url("tickets/modify/" . $ticket['id']);
             $buttonView = base_url("tickets/" . $ticket['id']);
-            $table->addRow(
-                // ["data" => $ticket['id'],"class"=>'p-5'],
-                explode("-", $ticket['id'])[4],
-                $ticket['tipus'],
-                ["data" =>  $ticket['descripcio'], "class" => " max-w-10 min-w-auto whitespace-nowrap overflow-hidden text-ellipsis"],
-                $ticket['emissor'],
-                ($ticket['receptor'] != lang('titles.toassign')) ? $ticket['receptor'] : lang('titles.toassign').' <i class="fa-solid fa-circle-exclamation text-xl text-red-600" ></i>',
 
-                date("d/m/Y", strtotime($ticket['created'])),
-                date("H:i", strtotime($ticket['created'])),
-                
-                ["data" => "<a class='flex p-3 justify-center  whitespace-nowrap w-full estat_" . $ticket['id_estat'] . "'>" . $ticket['estat'] . "</a>", "class" => "p-2 "],
-
-                [
-                    "data" =>
-                    "<a href='$buttonView' style='view-transition-name: info" . $ticket['id'] . ";' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-primario hover:text-secundario rounded-xl hover:rounded-xl transition-all ease-out duration-250 hover:transition hover:ease-in hover:duration-250 fa-eye'></i></a>
-                     <a href='$buttonUpdate' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-orange-600 hover:text-secundario hover:rounded-xl transition-all ease-out duration-250  rounded-xl hover:transition hover:ease-in hover:duration-250 fa-pencil'></i></a>
-                     <a href='$buttonDelete' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-red-800 hover:text-secundario hover:rounded-xl transition-all ease-out duration-250  rounded-xl hover:transition hover:ease-in hover:duration-250 fa-trash'></i></a>",
-
-                    "class" => " p-5 flex h-16 justify-between items-center"
-                ],
-
-            );
+            if (($role == "admin" ) || ($role== "sstt") || (($role=="ins") && ($ticket['id_estat']==1)) ) {
+                $table->addRow(
+                    // ["data" => $ticket['id'],"class"=>'p-5'],
+                    explode("-", $ticket['id'])[4],
+                    $ticket['tipus'],
+                    ["data" =>  $ticket['descripcio'], "class" => " max-w-10 min-w-auto whitespace-nowrap overflow-hidden text-ellipsis"],
+                    $ticket['emissor'],
+                    ($ticket['receptor'] != lang('titles.toassign')) ? $ticket['receptor'] : lang('titles.toassign').' <i class="fa-solid fa-circle-exclamation text-xl text-red-600" ></i>',
+    
+                    date("d/m/Y", strtotime($ticket['created'])),
+                    date("H:i", strtotime($ticket['created'])),
+                    
+                    ["data" => "<a class='flex p-3 justify-center  whitespace-nowrap w-full estat_" . $ticket['id_estat'] . "'>" . $ticket['estat'] . "</a>", "class" => "p-2 "],
+                    
+                    [
+                        "data" =>
+                        "<a href='$buttonView' style='view-transition-name: info" . $ticket['id'] . ";' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-primario hover:text-secundario rounded-xl hover:rounded-xl transition-all ease-out duration-250 hover:transition hover:ease-in hover:duration-250 fa-eye'></i></a>
+                         <a href='$buttonUpdate' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-orange-600 hover:text-secundario hover:rounded-xl transition-all ease-out duration-250  rounded-xl hover:transition hover:ease-in hover:duration-250 fa-pencil'></i></a>
+                         <a href='$buttonDelete' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-red-800 hover:text-secundario hover:rounded-xl transition-all ease-out duration-250  rounded-xl hover:transition hover:ease-in hover:duration-250 fa-trash'></i></a>",
+    
+                        "class" => " p-5 flex h-16 justify-between items-center"
+                    ],
+    
+                );
+            }elseif (("ins"==$role) && ($ticket['id_estat']!=1)) {
+                $table->addRow(
+                    // ["data" => $ticket['id'],"class"=>'p-5'],
+                    explode("-", $ticket['id'])[4],
+                    $ticket['tipus'],
+                    ["data" =>  $ticket['descripcio'], "class" => " max-w-10 min-w-auto whitespace-nowrap overflow-hidden text-ellipsis"],
+                    $ticket['emissor'],
+                    ($ticket['receptor'] != lang('titles.toassign')) ? $ticket['receptor'] : lang('titles.toassign').' <i class="fa-solid fa-circle-exclamation text-xl text-red-600" ></i>',
+    
+                    date("d/m/Y", strtotime($ticket['created'])),
+                    date("H:i", strtotime($ticket['created'])),
+                    
+                    ["data" => "<a class='flex p-3 justify-center  whitespace-nowrap w-full estat_" . $ticket['id_estat'] . "'>" . $ticket['estat'] . "</a>", "class" => "p-2 "],
+                    
+                    [
+                        "data" =>
+                        "<a href='$buttonView' style='view-transition-name: info" . $ticket['id'] . ";' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-primario hover:text-secundario rounded-xl hover:rounded-xl transition-all ease-out duration-250 hover:transition hover:ease-in hover:duration-250 fa-eye'></i></a>
+                         <a href='$buttonUpdate' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-orange-600 hover:text-secundario hover:rounded-xl transition-all ease-out duration-250  rounded-xl hover:transition hover:ease-in hover:duration-250 fa-pencil'></i></a>",
+    
+                        "class" => " p-5 flex h-16 justify-between items-center"
+                    ],
+    
+                );            
+            }else{
+                $table->addRow(
+                    // ["data" => $ticket['id'],"class"=>'p-5'],
+                    explode("-", $ticket['id'])[4],
+                    $ticket['tipus'],
+                    ["data" =>  $ticket['descripcio'], "class" => " max-w-10 min-w-auto whitespace-nowrap overflow-hidden text-ellipsis"],
+                    $ticket['emissor'],
+                    ($ticket['receptor'] != lang('titles.toassign')) ? $ticket['receptor'] : lang('titles.toassign').' <i class="fa-solid fa-circle-exclamation text-xl text-red-600" ></i>',
+    
+                    date("d/m/Y", strtotime($ticket['created'])),
+                    date("H:i", strtotime($ticket['created'])),
+                    
+                    ["data" => "<a class='flex p-3 justify-center  whitespace-nowrap w-full estat_" . $ticket['id_estat'] . "'>" . $ticket['estat'] . "</a>", "class" => "p-2 "],
+                    
+                    [
+                        "data" =>
+                        "<a href='$buttonView' style='view-transition-name: info" . $ticket['id'] . ";' class='p-2 btn btn-primary'><i class='fa-solid p-3 text-xl text-terciario-1 hover:bg-primario hover:text-secundario rounded-xl hover:rounded-xl transition-all ease-out duration-250 hover:transition hover:ease-in hover:duration-250 fa-eye'></i></a>",
+    
+                        "class" => " p-5 flex h-16 justify-between items-center"
+                    ],
+    
+                );   
+            }
 
             $count++;
         }
@@ -270,6 +321,9 @@ class TicketsController extends BaseController
     public function modifyTicket($id)
     {
         helper('form');
+        if (session()->get('user')['role']=="alumn" || session()->get('user')['role']=="prof") {
+            return redirect()->back();
+        }
 
         $type = new TipusDispositiuModel();
         $center = new CentreModel();
@@ -290,7 +344,9 @@ class TicketsController extends BaseController
 
         $model = new TiquetModel();
         helper('form');
-
+        if (session()->get('user')['role']=="alumn" || session()->get('user')['role']=="prof") {
+            return redirect()->back();
+        }
         $validationRules =
         [
             'description' => [
